requirements.txt
discord.py @ git+https://github.com/rapptz/discord.py@c1c64575985de28aa1d373e14a313e54c7786baf 
motor
dnspython
python-dotenv
requests
pymongo